Fix test
by extracting NewPhraseConfirm.hlpers|test|types - needed to avoid
issues of jest + !file-loader! definitions of icons, which were imported into
`NewPhraseConfirm` before
