+++ src/renderer/components/wallet/phrase/NewPhraseConfirm.tsx
@@ -12,45 +12,8 @@ import * as walletRoutes from '../../../routes/wallet'
 import { Button } from '../../uielements/button'
 import { Phrase } from './index'
 import * as NewPhraseStyled from './NewPhrase.styles'
+import { checkPhraseConfirmWordsFactory } from './NewPhraseConfirm.helper'
 import * as PhraseStyled from './Phrase.styles'
-
-export type WordType = {
-  text: string
-  _id: string
-  sequence?: number
-  error?: boolean
-  selected?: boolean
-}
-
-export const checkPhraseConfirmWordsFactory =
-  (setWordsList: (words: WordType[]) => void, setMnemonicError: (error: string) => void) =>
-  (words: WordType[], selectedWords: WordType[]) => {
-    if (words.length === selectedWords.length) {
-      let isErr = false
-      let newWords = [...words]
-
-      for (let i = 0; i < words.length; i++) {
-        const word = words[i]
-        const selectedWord = selectedWords[i]
-
-        if (word._id !== selectedWord._id) {
-          // eslint-disable-next-line no-loop-func
-          newWords = words.map((e: WordType) => {
-            if (e._id === selectedWord._id) {
-              e.error = true
-            }
-            isErr = true
-            return e
-          })
-        }
-      }
-      setWordsList(newWords)
-      return !isErr
-    } else {
-      setMnemonicError('Complete confirmation')
-      return false
-    }
-  }
 
 export const NewPhraseConfirm: React.FC<{ mnemonic: string; onConfirm: () => Promise<void> }> = ({
   mnemonic,
