+++ src/renderer/components/wallet/phrase/NewPhraseConfirm.types.ts
@@ -1,0 +1,7 @@
+export type WordType = {
+  text: string
+  _id: string
+  sequence?: number
+  error?: boolean
+  selected?: boolean
+}
